[MOV] (_ => sale_)loyalty_delivery: merge modules
There is no need to have two bridges linking delivery & loyalty.
The two modules held the two parts of the same logic, making no
sense to keep the two modules separate.

Task-3420579
