--- addons/loyalty_delivery/i18n/loyalty_delivery.pot
@@ -1,69 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* loyalty_delivery
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server saas~16.3\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2023-05-30 05:09+0000\n"
-"PO-Revision-Date: 2023-05-30 05:09+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: loyalty_delivery
-#. odoo-python
-#: code:addons/loyalty_delivery/models/loyalty_reward.py:0
-#, python-format
-msgid " (Max %s)"
-msgstr ""
-
-#. module: loyalty_delivery
-#: model_terms:ir.ui.view,arch_db:loyalty_delivery.loyalty_reward_view_kanban_inherit_loyalty_delivery
-msgid "( Max"
-msgstr ""
-
-#. module: loyalty_delivery
-#. odoo-python
-#: code:addons/loyalty_delivery/models/loyalty_program.py:0
-#, python-format
-msgid "Automatic promotion: free shipping on orders higher than $50"
-msgstr ""
-
-#. module: loyalty_delivery
-#: model:ir.model.fields.selection,name:loyalty_delivery.selection__loyalty_reward__reward_type__shipping
-msgid "Free Shipping"
-msgstr ""
-
-#. module: loyalty_delivery
-#. odoo-python
-#: code:addons/loyalty_delivery/models/loyalty_reward.py:0
-#: model_terms:ir.ui.view,arch_db:loyalty_delivery.loyalty_reward_view_form_inherit_loyalty_delivery
-#: model_terms:ir.ui.view,arch_db:loyalty_delivery.loyalty_reward_view_kanban_inherit_loyalty_delivery
-#, python-format
-msgid "Free shipping"
-msgstr ""
-
-#. module: loyalty_delivery
-#: model:ir.model,name:loyalty_delivery.model_loyalty_program
-msgid "Loyalty Program"
-msgstr ""
-
-#. module: loyalty_delivery
-#: model:ir.model,name:loyalty_delivery.model_loyalty_reward
-msgid "Loyalty Reward"
-msgstr ""
-
-#. module: loyalty_delivery
-#: model:ir.model.fields,field_description:loyalty_delivery.field_loyalty_reward__reward_type
-msgid "Reward Type"
-msgstr ""
-
-#. module: loyalty_delivery
-#: model:ir.model,name:loyalty_delivery.model_sale_order
-msgid "Sales Order"
-msgstr ""
