--- addons/loyalty_delivery/models/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import loyalty_program
-from . import loyalty_reward
-from . import sale_order
