+++ addons/sale_loyalty_delivery/__manifest__.py
@@ -5,8 +5,10 @@
     'summary': 'Adds free shipping mechanism in sales orders',
     'description': 'Integrate free shipping in sales orders.',
     'category': 'Sales/Sales',
-    'version': '1.0',
-    'depends': ['sale_loyalty', 'loyalty_delivery'],
+    'data': [
+        'views/loyalty_reward_views.xml',
+    ],
+    'depends': ['sale_loyalty', 'delivery'],
     'auto_install': True,
     'license': 'LGPL-3',
 }
