+++ addons/sale_loyalty_delivery/models/__init__.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import loyalty_program
+from . import loyalty_reward
 from . import sale_order
 from . import sale_order_line
