+++ addons/sale_loyalty_delivery/tests/__init__.py
@@ -2,3 +2,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import test_free_shipping_reward
+from . import test_loyalty_delivery
