+++ addons/sale_loyalty_delivery/tests/test_free_shipping_reward.py
@@ -10,7 +10,7 @@ class TestSaleCouponProgramRules(TestSaleCouponCommon):
 
     @classmethod
     def setUpClass(cls):
-        super(TestSaleCouponProgramRules, cls).setUpClass()
+        super().setUpClass()
         cls.iPadMini = cls.env['product.product'].create({'name': 'Large Cabinet', 'list_price': 320.0})
         tax_15pc_excl = cls.env['account.tax'].create({
             'name': "15% Tax excl",
