+++ addons/sale_loyalty_delivery/views/loyalty_reward_views.xml
@@ -1,15 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
+
     <record id="loyalty_reward_view_form_inherit_loyalty_delivery" model="ir.ui.view">
         <field name="name">loyalty.reward.view.form.inherit.loyalty.delivery</field>
         <field name="model">loyalty.reward</field>
         <field name="inherit_id" ref="loyalty.loyalty_reward_view_form"/>
         <field name="arch" type="xml">
-            <xpath expr="//group[@id='reward_type_group']" position="after">
+            <group name="reward_type_group" position="after">
                 <group id="shipping" string="Free shipping" attrs="{'invisible': [('reward_type', '!=', 'shipping')]}">
                     <field name="discount_max_amount"/>
                 </group>
-            </xpath>
+            </group>
         </field>
     </record>
 
@@ -18,13 +19,14 @@
         <field name="model">loyalty.reward</field>
         <field name="inherit_id" ref="loyalty.loyalty_reward_view_kanban"/>
         <field name="arch" type="xml">
-            <xpath expr="//div[@id='reward_info']" position="inside">
-                <t t-if="record.reward_type.raw_value === 'shipping'">
-
+            <div name="reward_info" position="inside">
+                <t t-elif="record.reward_type.raw_value === 'shipping'">
                     <a>Free shipping <t t-if="record.discount_max_amount.raw_value > 0">( Max <field name="discount_max_amount"/> )</t></a>
-                    <br/><br/>
-                </t>                    
-            </xpath>
+                    <br/>
+                    <br/>
+                </t>
+            </div>
         </field>
     </record>
+
 </odoo>
